# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Service culturel - Plateforme web API",
    "description": "API de gestion des inscriptions scolaires aux événements de l'Opéra Orchestre National Montpellier Occitanie.",
    "version": "1.6.7",
    "contact": {
      "name": "Service culturel - Plateforme web",
      "url": "https://github.com/vincbct34/service-culturel-plateforme-web"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Serveur courant"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token JWT (header Authorization: Bearer <token>)"
      },
      "CronAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "CRON_SECRET token pour les tâches planifiées"
      },
      "CSRFToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Jeton CSRF requis pour POST/PUT/PATCH/DELETE authentifiés"
      }
    },
    "schemas": {
...
      "PublicCategory": {
        "type": "string",
        "enum": [
          "CRECHE",
          "MATERNELLE",
          "ELEMENTAIRE",
          "COLLEGE",
          "LYCEE",
          "SUPERIEUR",
          "CONSERVATOIRE",
          "ASSOCIATION",
          "PERISCOLAIRE",
          "PUBLICS_EMPECHES",
          "AUTRE"
        ]
      },
      "SchoolGrade": {
        "type": "string",
        "enum": [
          "PS",
          "MS",
          "GS",
          "CP",
          "CE1",
          "CE2",
          "CM1",
          "CM2",
          "SIXIEME",
          "CINQUIEME",
          "QUATRIEME",
          "TROISIEME",
          "SECONDE",
          "PREMIERE",
          "TERMINALE"
        ]
      },
      "AgeRange": {
        "type": "string",
        "enum": ["AGE_0_3", "AGE_3_6", "AGE_6_11", "AGE_11_15", "AGE_15_18", "AGE_18_PLUS"]
      },
      "EventType": {
        "type": "string",
        "enum": [
          "OPERA",
          "CONCERT_LYRIQUE",
          "SYMPHONIQUE",
          "CHAMBRE_BAROQUE",
          "EN_FAMILLE",
          "OPERA_JUNIOR",
          "CINE_CONCERT",
          "INSOLITE",
          "THEATRE_MUSICAL",
          "DANSE",
          "CONTE_MUSICAL",
          "CONCERT_DECENTRALISE",
          "ELECTRO_ACOUSTIQUE",
          "MUSIQUE_ELECTRONIQUE",
          "CONCERT_LECTURE",
          "PLEIN_AIR",
          "JAZZ",
          "LITTERATURE",
          "MASTERCLASS",
          "MUSIQUE_ET_BIEN_ETRE",
          "MUSIQUE_ACTUELLE",
          "PARTICIPATIF",
          "PROMENADE_SONORE",
          "ATELIER",
          "GALA",
          "EXPOSITION",
          "PORTES_OUVERTES",
          "INTERDISCIPLINAIRE",
          "SPECTACLE_EDUCATIF",
          "CARITATIF",
          "THEATRE",
          "RENDEZ_VOUS",
          "MUSIQUES_DAILLEURS",
          "BAROQUE"
        ]
      },
      "EventStatus": {
        "type": "string",
        "enum": ["OPEN", "CLOSED", "ARCHIVED"]
      },
      "Accessibility": {
        "type": "string",
        "enum": ["VISUAL", "AUDITORY", "MOTOR", "PSYCHIC", "NEUROATYPICAL", "OTHER"]
      },
...
      "Event": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventType"
            }
          },
          "category": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicCategory"
            }
          },
          "grades": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SchoolGrade"
            }
          },
          "age_ranges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgeRange"
            }
          },
          "location": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "total_seats": {
            "type": "integer"
          },
          "booked_seats": {
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          },
          "image_url": {
            "type": "string",
            "nullable": true
          },
          "event_dates": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            }
          },
          "accessibility": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Accessibility"
            }
          },
          "has_initial_formation": {
            "type": "boolean"
          },
          "is_formation_mandatory": {
            "type": "boolean"
          },
          "has_musical_preparation": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/admin/events/{id}": {
      "get": {
        "tags": ["Admin – Events"],
        "summary": "Détail d'un événement (admin)",
        "description": "Retourne les informations détaillées d'un événement avec accessibilité et compteurs.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détail de l'événement",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "event": {
                      "$ref": "#/components/schemas/Event"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Événement non trouvé"
          }
        }
      },
      "put": {
        "tags": ["Admin – Events"],
        "summary": "Modifier un événement",
        "description": "Met à jour un événement. Supporte les mises à jour partielles. Gère les protected_fields vis-à-vis du scraper.",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CSRFToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "type": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/EventType"
                    }
                  },
                  "location": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "number"
                  },
                  "total_seats": {
                    "type": "integer"
                  },
                  "status": {
                    "$ref": "#/components/schemas/EventStatus"
                  },
                  "image_url": {
                    "type": "string"
                  },
                  "event_dates": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "category": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PublicCategory"
                    }
                  },
                  "grades": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/SchoolGrade"
                    }
                  },
                  "age_ranges": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/AgeRange"
                    }
                  },
                  "has_initial_formation": {
                    "type": "boolean"
                  },
                  "is_formation_mandatory": {
                    "type": "boolean"
                  },
                  "has_musical_preparation": {
                    "type": "boolean"
                  },
                  "accessibility": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "slug": {
                    "type": "string"
                  },
                  "protected_fields": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Événement mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "event": {
                      "$ref": "#/components/schemas/Event"
                    },
                    "message": {
                      "type": "string"
                    },
                    "modified_fields": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "protected_fields": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Admin – Events"],
        "summary": "Supprimer un événement",
        "description": "Supprime un événement. Bloqué si l'événement a des inscriptions.",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CSRFToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Événement supprimé",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Événement supprimé avec succès"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Événement avec inscriptions"
          }
        }
      }
    },
...
  }
}